\name{deg}    
\title{Degrees}
\alias{deg}

\description{
Converts radians to degrees. 
}

\usage{
deg(radian)
}

\arguments{
\item{radian}{vector of radian measurements.}
}

\value{
Returns a vector of degree measurements corresponding to the data in vector radian.
}

\seealso{
\link{rad}
}

\keyword{math}
