\name{I.0}
\title{Zeroth Order Bessel Function of the First Kind}
\alias{I.0}

\description{
An alias of besselI(x, nu=0, expon.scaled = FALSE) used for compatibility with old version package code.
}

\usage{
I.0(x)
}

\arguments{
\item{x}{non-negative numerical value at which to evaluate the Bessel function.}
}

\value{
Returns the zeroth order Bessel function of the first kind evaluated at a specified real number.
}

\details{
The use of this function is deprecated. Please use \code{\link{besselI}}.
}

\seealso{
\code{\link{besselI}}.
}

\keyword{math}

