\name{I.p}
\title{P-th Order Bessel Function of the First Kind}
\alias{I.p}

\description{
An alias of besselI(x, nu=p, expon.scaled = FALSE) used for compatibility with old version package code.
}

\usage{
I.p(p, x)
}

\arguments{
\item{p}{positive integer order of the Bessel function.}
\item{x}{non-negative numerical value at which to evaluate the Bessel function.}
}

\value{
Returns the p-th order Bessel function of the first kind, evaluated at a specified real number.
}

\details{
The use of this function is deprecated. Please use \code{\link{besselI}}.
}


\seealso{
\code{\link{besselI}}.
}

\keyword{math}

