\name{rad}    
\title{Radians}
\alias{rad}

\description{
Converts degrees to radians. 
}

\usage{
rad(degree)
}

\arguments{
\item{degree}{vector of degree measurements.}
}

\value{
Returns a vector of radian measurements corresponding to the data in vector degree.
}

\seealso{
\link{deg}
}

\keyword{math}


